fix: resolve lint findings in taskresource (gci, misspell, unconvert)
golangci-lint run ./internal/util/taskresource/... reported 12 issues (5
misspell, 2 gci, 5 unconvert). Code Check runs before golangci-lint in
this repo's CI, so this alone would have blocked the PR before any real
linting ran.

- gci: request.go and request_test.go each split the
  github.com/milvus-io/milvus-proto and github.com/milvus-io/milvus
  imports into two blank-line-separated groups; this repo's gci config
  (sections: standard, default, prefix(github.com/milvus-io)) puts both
  under one prefix group. Auto-fixed via `golangci-lint run --fix`.

- misspell: unrecognised/recognised/recognise -> unrecognized/recognized
  /recognize (UK -> US spelling, matching this sweep's convention
  elsewhere: serialise->serialize, behaviour->behavior, licence->license).
  Comment text only, including one instance in estimate_index.go (a
  pre-existing Task 1-4 file in the same package).

- unconvert: five `int64(128*mib)`-shaped assertions in
  estimate_import_test.go. Checked each individually before removing:
  mib is declared `const mib = int64(1) << 20`, an explicitly int64-typed
  constant, so `128*mib` is already int64 and the wrapping int64(...) is
  redundant on both sides of the comparison. All five were test-assertion
  literals with no load-bearing type change; none were left in place.

No function signature, exported name, or runtime behavior changed in
request.go; the diff there is an import-group merge and one comment fix.
Full package suite (63 tests, including pre-existing Task 1-6 coverage)
still passes at 98.5% coverage; golangci-lint run ./internal/util/task
resource/... now reports 0 issues.

+++ internal/util/taskresource/estimate_import_test.go
@@ -143,7 +143,7 @@ func TestEstimateImportExactValueForCommonCase(t *testing.T) {
 	// perFile = 16MiB base * 2 vchannels * 1 partition = 32MiB.
 	got := EstimateImport(ImportInput{FileNum: 4, VChannelNum: 2, PartitionNum: 1})
 
-	assert.Equal(t, int64(128*mib), got.Memory)
+	assert.Equal(t, 128*mib, got.Memory)
 	assert.Equal(t, 2.0, got.CPU)
 }
 
@@ -163,7 +163,7 @@ func TestEstimateImportL0UsesDeleteBufferExactValue(t *testing.T) {
 
 	deleteBuffer := pt.DataNodeCfg.ImportDeleteBufferSize.GetAsInt64()
 	assert.Equal(t, deleteBuffer*8, got.Memory)
-	assert.Equal(t, int64(128*mib), got.Memory)
+	assert.Equal(t, 128*mib, got.Memory)
 }
 
 func TestEstimatePreImportExactValue(t *testing.T) {
@@ -181,7 +181,7 @@ func TestEstimatePreImportExactValue(t *testing.T) {
 
 	baseBuffer := pt.DataNodeCfg.ImportBaseBufferSize.GetAsInt64()
 	assert.Equal(t, baseBuffer*8, got.Memory)
-	assert.Equal(t, int64(128*mib), got.Memory)
+	assert.Equal(t, 128*mib, got.Memory)
 }
 
 // A zero VChannelNum/PartitionNum must be floored to 1, not treated as a
@@ -205,7 +205,7 @@ func TestEstimateImportZeroFanoutFloorsToOne(t *testing.T) {
 	one := EstimateImport(ImportInput{FileNum: 8, VChannelNum: 1, PartitionNum: 1})
 
 	assert.Equal(t, one.Memory, zero.Memory)
-	assert.Equal(t, int64(128*mib), zero.Memory)
+	assert.Equal(t, 128*mib, zero.Memory)
 }
 
 // MaxFileMemorySize must cap the per-file buffer before it is multiplied by
@@ -229,7 +229,7 @@ func TestEstimateImportCapsAtMaxFileMemorySize(t *testing.T) {
 		MaxFileMemorySize: 20 * mib,
 	})
 
-	assert.Equal(t, int64(80*mib), got.Memory)
+	assert.Equal(t, 80*mib, got.Memory)
 	assert.Equal(t, 2.0, got.CPU)
 }
